+++ data_generator/devops/devops_nginx.go
@@ -83,7 +83,7 @@ func (m *NginxMeasurement) ToPoint(p *Point) bool {
 		}
 		idx := cache & letterIdxMask
 		// value := atomic.AddInt64(&m.fieldValues[i], idx)
-		p.AppendField(NginxFields[i].Label, idx) // 0~128之间随机整数
+		p.AppendField(NginxFields[i].Label, int64(idx)) // 0~128之间随机整数
 		i--
 
 		cache >>= letterIdxBits
