+++ data_generator/devops/devops_redis.go
@@ -122,7 +122,7 @@ func (m *RedisMeasurement) ToPoint(p *Point) bool {
 		}
 		idx := cache & letterIdxMask
 		// value := atomic.AddInt64(&m.fieldValues[i], idx)
-		p.AppendField(RedisFields[i].Label, idx) // 0~1024之间随机整数
+		p.AppendField(RedisFields[i].Label, int64(idx)) // 0~1024之间随机整数
 		i--
 
 		cache >>= letterIdxBits
